Add structure and make more idiomatic
