+++ src/lib.rs
@@ -1,1 +1,10 @@
+mod constants;
+mod errors;
 pub mod ratchet;
+mod utils;
+
+#[cfg(test)]
+mod tests;
+
+pub use errors::*;
+pub use ratchet::Ratchet;
